+++ microsphere-configuration-etcd-spring/src/main/java/io/microsphere/configuration/nacos/spring/annotation/EtcdPropertySourceLoader.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.microsphere.configuration.etcd.spring.annotation;
+package io.microsphere.configuration.nacos.spring.annotation;
 
 import io.etcd.jetcd.ByteSequence;
 import io.etcd.jetcd.Client;
