+++ microsphere-configuration-etcd-spring/src/test/java/io/microsphere/configuration/nacos/spring/annotation/EtcdPropertySourceTest.java
@@ -14,14 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.microsphere.configuration.etcd.spring.annotation;
+package io.microsphere.configuration.nacos.spring.annotation;
 
-import io.microsphere.nacos.client.NacosClientConfig;
-import io.microsphere.nacos.client.OpenApiVersion;
-import io.microsphere.nacos.client.common.OpenApiTemplateClient;
-import io.microsphere.nacos.client.common.config.ConfigClient;
-import io.microsphere.nacos.client.v1.OpenApiNacosClient;
-import io.microsphere.nacos.client.v2.OpenApiNacosClientV2;
+import io.etcd.jetcd.ByteSequence;
+import io.etcd.jetcd.Client;
+import io.etcd.jetcd.KV;
 import io.microsphere.spring.config.env.support.JsonPropertySourceFactory;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -36,30 +33,32 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.util.StreamUtils;
 
+import java.nio.charset.StandardCharsets;
+
 import static org.junit.Assert.assertEquals;
 
 /**
- * {@link NacosPorpertySource} Test
+ * {@link EtcdPropertySource} Test
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since 1.0.0
  */
 @RunWith(SpringRunner.class)
 @ContextConfiguration(classes = {
-        NacosPorpertySourceTest.class,
-        NacosPorpertySourceTest.Config.class
+        EtcdPropertySourceTest.class,
+        EtcdPropertySourceTest.Config.class
 })
-public class NacosPorpertySourceTest {
+public class EtcdPropertySourceTest {
 
     @Autowired
     private Environment environment;
 
-    private static ConfigClient client;
+    private static Client client;
 
     @BeforeClass
     public static void init() throws Exception {
-        NacosPorpertySource annotation =
-                NacosPorpertySourceTest.Config.class.getAnnotation(NacosPorpertySource.class);
+        EtcdPropertySource annotation =
+                EtcdPropertySourceTest.Config.class.getAnnotation(EtcdPropertySource.class);
 
         client = buildClient(annotation);
 
@@ -67,59 +66,56 @@ public class NacosPorpertySourceTest {
         mockConfig();
     }
 
-    private static ConfigClient buildClient(NacosPorpertySource annotation) throws Exception {
-        NacosClientConfig config = new NacosClientConfig();
-        config.setServerAddress(annotation.serverAddress());
-        if (OpenApiVersion.V1.equals(annotation.openApiVersion())) {
-            return new OpenApiNacosClient(config);
-        } else if (OpenApiVersion.V2.equals(annotation.openApiVersion())) {
-            return new OpenApiNacosClientV2(config);
-        } else {
-            throw new RuntimeException("Unsupported nacos open api version " + annotation.openApiVersion());
-        }
+    private static Client buildClient(EtcdPropertySource annotation) throws Exception {
+        Client client = Client.builder()
+                .endpoints(annotation.endpoints())
+                .build();
+        return client;
     }
 
     private static void mockConfig() throws Exception {
 
         ResourcePatternResolver patternResolver = new PathMatchingResourcePatternResolver();
 
-        Resource[] resources = patternResolver.getResources("classpath:/META-INF/nacos/*.json");
+        Resource[] resources = patternResolver.getResources("classpath:/META-INF/etcd/*.json");
 
         for (Resource resource : resources) {
             // test.json
             String fileName = resource.getFilename();
             String key = fileName;
             byte[] data = StreamUtils.copyToByteArray(resource.getInputStream());
-            writeConfig(key, new String(data));
+            writeConfig(key, data);
         }
     }
 
-    private static void writeConfig(String stringKey, String data) throws Exception {
-        client.publishConfigContent("DEFAULT_GROUP", stringKey, data);
+    private static void writeConfig(String stringKey, byte[] data) throws Exception {
+        KV kvClient = client.getKVClient();
+        ByteSequence key = ByteSequence.from(stringKey.getBytes());
+        ByteSequence value = ByteSequence.from(data);
+        // put the key-value
+        kvClient.put(key, value).get();
     }
 
     @AfterClass
-    public static void destroy() throws Exception {
-        ((OpenApiTemplateClient) client).getOpenApiClient().close();
+    public static void destroy() {
+        client.close();
     }
 
 
     @Test
     public void test() throws Exception {
-        assertEquals("microsphere", environment.getProperty("my.name"));
+        assertEquals("mercyblitz", environment.getProperty("my.name"));
 
-        writeConfig("test.json", "{ \"my.name\": \"Microsphere\" }");
+        writeConfig("test.json", "{ \"my.name\": \"Mercy Ma\" }".getBytes(StandardCharsets.UTF_8));
 
         Thread.sleep(5 * 1000);
 
-        assertEquals("Microsphere", environment.getProperty("my.name"));
+        assertEquals("Mercy Ma", environment.getProperty("my.name"));
     }
 
-    @NacosPorpertySource(
+    @EtcdPropertySource(
             key = "test.json",
-            factory = JsonPropertySourceFactory.class,
-            serverAddress = "",
-            openApiVersion = OpenApiVersion.V2)
+            factory = JsonPropertySourceFactory.class)
     static class Config {
 
     }
