+++ microsphere-configuration-nacos-spring/src/main/java/io/microsphere/configuration/nacos/spring/annotation/NacosPropertySourceAttributes.java
@@ -14,24 +14,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.microsphere.configuration.etcd.spring.annotation;
+package io.microsphere.configuration.nacos.spring.annotation;
 
+import io.microsphere.nacos.client.OpenApiVersion;
 import io.microsphere.spring.config.context.annotation.PropertySourceExtensionAttributes;
 import org.springframework.core.env.PropertyResolver;
 
 import java.util.Map;
 
 /**
- * The {@link PropertySourceExtensionAttributes} for {@link EtcdPropertySource}
+ * The {@link PropertySourceExtensionAttributes} for {@link NacosPorpertySource}
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
- * @see EtcdPropertySource
+ * @see NacosPorpertySource
  * @see PropertySourceExtensionAttributes
  * @since 1.0.0
  */
-public class EtcdPropertySourceAttributes extends PropertySourceExtensionAttributes<EtcdPropertySource> {
+public class NacosPropertySourceAttributes extends PropertySourceExtensionAttributes<NacosPorpertySource> {
 
-    public EtcdPropertySourceAttributes(Map<String, Object> another, Class<EtcdPropertySource> annotationType, PropertyResolver propertyResolver) {
+    public NacosPropertySourceAttributes(Map<String, Object> another, Class<NacosPorpertySource> annotationType, PropertyResolver propertyResolver) {
         super(another, annotationType, propertyResolver);
     }
 
@@ -39,11 +40,11 @@ public class EtcdPropertySourceAttributes extends PropertySourceExtensionAttribu
         return getValue();
     }
 
-    public final String getTarget() {
-        return getString("target");
+    public final OpenApiVersion getVersion() {
+        return getEnum("openApiVersion");
     }
 
-    public final String[] getEndpoints() {
-        return getStringArray("endpoints");
+    public final String getServerAddress() {
+        return getString("serverAddress");
     }
 }
